Basic index page and Make changes
A few new files here. We are creating a bunch of basic HTML templates
using the Bootstrap framework. This has given us a basic index page with
a nav bar and empty footer.

Slight changes to Makefile to allow for quicker testing locally. We
don't need to build a new container every time.